tests/test_state_management.py
```python
# Importing necessary libraries
import unittest
import tensorflow as tf
from tensorflow.keras import Sequential
from tensorflow.keras.layers import Dense
from src.state_management import StateManagement
import os

class TestStateManagement(unittest.TestCase):
    def setUp(self):
        # Create a simple model for testing
        self.model = Sequential([Dense(units=1, input_shape=[1])])
        self.model.compile(optimizer='sgd', loss='mean_squared_error')

        # Create an instance of the StateManagement class
        self.state_management = StateManagement(self.model)

        # Create directories for testing
        self.checkpoint_dir = './checkpoints'
        self.model_dir = './models'
        self.snapshot_dir = './snapshots'
        self.old_model_path = './old_model'
        self.new_model_path = './new_model'
        os.makedirs(self.checkpoint_dir, exist_ok=True)
        os.makedirs(self.model_dir, exist_ok=True)
        os.makedirs(self.snapshot_dir, exist_ok=True)
        os.makedirs(self.old_model_path, exist_ok=True)
        os.makedirs(self.new_model_path, exist_ok=True)

    def test_model_checkpointing(self):
        self.state_management.model_checkpointing(self.checkpoint_dir)
        self.assertTrue(os.listdir(self.checkpoint_dir))

    def test_model_version_control(self):
        self.state_management.model_version_control(self.model_dir, '1')
        self.assertTrue(os.path.exists(f"{self.model_dir}/1"))

    def test_model_snapshot(self):
        self.state_management.model_snapshot(self.snapshot_dir)
        self.assertTrue(os.path.exists(f"{self.snapshot_dir}/my_model.h5"))

    def test_model_migration(self):
        # Save a model to the old model path
        self.model.save(f"{self.old_model_path}/my_model.h5")
        self.state_management.model_migration(f"{self.old_model_path}/my_model.h5", self.new_model_path)
        self.assertTrue(os.path.exists(f"{self.new_model_path}/my_model.h5"))

if __name__ == '__main__':
    unittest.main()
```
